examples/langchain-cli/requirements.txt
langchain>=1.0
langchain-mcp-adapters
langchain-openai
requests
python-dotenv
rich

